refactor device functions out of simulation commands

+++ crates/spicy_simulate/src/solver/klu/solve.rs
@@ -374,7 +374,7 @@ pub fn solve(
                             let end = offp[k + 1];
                             temp[0] = x[k];
                             for p in start..end {
-                                let i = offi[p] as usize;
+                                let i = offi[p];
                                 x[i] -= offx[p] * temp[0];
                             }
                         }
@@ -386,7 +386,7 @@ pub fn solve(
                             temp[0] = x[2 * k];
                             temp[1] = x[2 * k + 1];
                             for p in start..end {
-                                let i = offi[p] as usize;
+                                let i = offi[p];
                                 x[2 * i] -= offx[p] * temp[0];
                                 x[2 * i + 1] -= offx[p] * temp[1];
                             }
@@ -400,7 +400,7 @@ pub fn solve(
                             temp[1] = x[3 * k + 1];
                             temp[2] = x[3 * k + 2];
                             for p in start..end {
-                                let i = offi[p] as usize;
+                                let i = offi[p];
                                 x[3 * i] -= offx[p] * temp[0];
                                 x[3 * i + 1] -= offx[p] * temp[1];
                                 x[3 * i + 2] -= offx[p] * temp[2];
@@ -416,7 +416,7 @@ pub fn solve(
                             temp[2] = x[4 * k + 2];
                             temp[3] = x[4 * k + 3];
                             for p in start..end {
-                                let i = offi[p] as usize;
+                                let i = offi[p];
                                 x[4 * i] -= offx[p] * temp[0];
                                 x[4 * i + 1] -= offx[p] * temp[1];
                                 x[4 * i + 2] -= offx[p] * temp[2];
